Fix subscription-renewal invoice product price issue.
Previously the product price would be the total price. Which was fine unless there's tax. In that case the product price would not be accurate and the total would be correct in terms of what was paid, but would be mathematically incorrect as the sum of the product and taxes.