Do not always re-enable soft-deleteable filter
This concerns the the "soft-deleteable" filter on the broadly used EcommerceEntityManager. If it is not enabled when paymentsRepository->getAllUsersPayments is called, then it should not be enabled at the end of that method.

Previously it was enabled at the end of that method. This was under the assumption that it was enabled when the method was called, and thus had been disabled. But if it already was disabled, then it did not need to be disabled, and thus on exit no "correcting manipulation" is required.